+++ front-end-tools/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix error message when specifying input parameter without uploading schema.
 - Fix error when contract calls return BigInts.
 - Fix contract method names in error messages.
+- Give warning when deploying a module that does not have embedded build information.
 
 ## 3.0.1